refactor(og-website): rename Open Graph lib folder
